requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==1.0.1
langchain-neo4j==0.5.0
python-dotenv==1.0.0
openai==1.14.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting
flake8==6.1.0
mypy==1.7.1
black==23.12.1

# Development
ipython==8.18.1
ipdb==0.13.13
